api for user management and upload log copilot

+++ migrations/20240710112230-create-user-bridev.js
@@ -23,7 +23,7 @@ module.exports = {
         unique: true,
       },
       status: {
-        type: Sequelize.TINYINT,
+        type: Sequelize.INTEGER,
         allowNull: false,
         defaultValue: 1
       },
